chore: Se han unificado los componentes de CommentsUserComponent y CommentsGameComponent en CommentsComponent para evitar duplicidad de código
