+++ src/components/Comments/CommentsComponent/index.tsx
@@ -2,12 +2,19 @@ import type { FC } from 'react';
 import React, { useEffect, useState } from 'react';
 import Error from '@/components/Error';
 import LoadingSpinner from '@/components/LoadingSpinner';
-import type { CommentsGame } from '@/models/Comment/types';
+import type { CommentsUser } from '@/models/Comment/types';
 import type { Game } from '@/models/Game/types';
+import type { User } from '@/models/User/types';
 import { getUserCookie } from '@/utils/login';
 import CommentsSection from '../CommentsSection';
 import InputFieldSection from '../InputFieldSection';
 import { Container } from './styles';
+
+export type CommentToPostUser = {
+  userWhoComments: number;
+  userCommented: number;
+  commentText: string;
+};
 
 export type CommentToPostGame = {
   userWhoComments: number;
@@ -16,17 +23,18 @@ export type CommentToPostGame = {
 };
 
 export type Props = {
-  gameData: Game;
+  data: User | Game;
+  type: string;
 };
 
-const CommentsGameContainer: FC<Props> = ({ gameData }) => {
+const CommentsComponent: FC<Props> = ({ data, type }) => {
   const user = getUserCookie();
-  const [commentsArray, setCommentsArray] = useState<CommentsGame[] | null>(null);
+  const [commentsArray, setCommentsArray] = useState<CommentsUser [] | null>(null);
   const [loading, setLoading] = useState<boolean>(false);
   const [error, setError] = useState<boolean>(false);
   const [post, setPost] = useState(false);
 
-  const postComment = async (commentToPost: CommentToPostGame) => {
+  const postComment = async (commentToPost: CommentToPostUser | CommentToPostGame) => {
     const url = 'https://gametrail-backend-production-8fc0.up.railway.app/api/comment';
 
     try {
@@ -35,9 +43,9 @@ const CommentsGameContainer: FC<Props> = ({ gameData }) => {
         headers: { Authorization: `Token ${user?.auth_token}`, 'Content-Type': 'application/json' },
         body: JSON.stringify(commentToPost),
       });
-      const data = await res.json();
+      const response = await res.json();
       // eslint-disable-next-line no-console
-      console.log({ data });
+      console.log({ response });
     } catch (err) {
       // eslint-disable-next-line no-console
       console.log({ err });
@@ -47,16 +55,39 @@ const CommentsGameContainer: FC<Props> = ({ gameData }) => {
   const onClickNewComment = (input: string) => {
     if (input === '' || input === undefined) return;
 
-    const commentToPost = {
-      userWhoComments: user?.id ?? 1,
-      game: gameData.id,
-      commentText: input,
-    };
+    let commentToPost;
+    if (type === 'user') {
+      commentToPost = {
+        userWhoComments: user?.id ?? 1,
+        userCommented: data.id,
+        commentText: input,
+      };
+    } else {
+      commentToPost = {
+        userWhoComments: user?.id ?? 1,
+        game: data.id,
+        commentText: input,
+      };
+    }
     postComment(commentToPost);
     setPost(!post);
   };
 
   useEffect(() => {
+    const fetchUser = async (id: number) => {
+      setLoading(true);
+      try {
+        const response = await fetch(`https://gametrail-backend-production-8fc0.up.railway.app/api/user/${id}/`);
+        const userData = await response.json();
+        setCommentsArray(userData.comments_received);
+        setError(false);
+      } catch (err) {
+        setError(true);
+      } finally {
+        setLoading(false);
+      }
+    };
+
     const fetchGameDetails = async (id: number) => {
       setLoading(true);
       try {
@@ -70,11 +101,15 @@ const CommentsGameContainer: FC<Props> = ({ gameData }) => {
         setLoading(false);
       }
     };
-    fetchGameDetails(gameData.id);
-  }, [post, gameData.id]);
+    if (type === 'user') {
+      fetchUser(data.id);
+    } else {
+      fetchGameDetails(data.id);
+    }
+  }, [post, type, data.id]);
 
   if (loading) return <LoadingSpinner />;
-  if (error || !gameData) return <Error />;
+  if (error || !data) return <Error />;
   return (
     <Container>
       <InputFieldSection onClickNewComment={onClickNewComment} />
@@ -82,4 +117,4 @@ const CommentsGameContainer: FC<Props> = ({ gameData }) => {
     </Container>
   );
 };
-export default CommentsGameContainer;
+export default CommentsComponent;
